Move the rest of the old dynamo package
