+++ sierra_adapter/sierra_items_to_dynamo/src/test/scala/uk/ac/wellcome/platform/sierra_items_to_dynamo/ServerTest.scala
@@ -5,9 +5,8 @@ import com.twitter.finagle.http.Status._
 import org.scalatest.FunSpec
 import uk.ac.wellcome.messaging.test.fixtures.SQS
 import uk.ac.wellcome.models.transformable.sierra.SierraItemRecord
+import uk.ac.wellcome.storage.dynamo._
 import uk.ac.wellcome.test.fixtures.LocalDynamoDb
-
-import uk.ac.wellcome.dynamo._
 
 class ServerTest
     extends FunSpec
