+++ sierra_adapter/sierra_items_to_dynamo/src/test/scala/uk/ac/wellcome/platform/sierra_items_to_dynamo/SierraItemsToDynamoFeatureTest.scala
@@ -8,13 +8,13 @@ import org.scalatest.concurrent.Eventually
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.utils.JsonUtil._
 import uk.ac.wellcome.test.utils.ExtendedPatience
-import uk.ac.wellcome.dynamo._
 import uk.ac.wellcome.messaging.sqs.SQSMessage
 import uk.ac.wellcome.messaging.test.fixtures.SQS
 import uk.ac.wellcome.models.transformable.sierra.{
   SierraItemRecord,
   SierraRecord
 }
+import uk.ac.wellcome.storage.dynamo._
 import uk.ac.wellcome.test.fixtures.LocalDynamoDb
 
 class SierraItemsToDynamoFeatureTest
